refactor(wrapper): refactor cache_generator_call
